Fix custom_metrics.py

+++ regression/steps/custom_metrics.py
@@ -21,8 +21,8 @@ from sklearn.metrics import mean_squared_error
 
 def weighted_mean_squared_error(
     eval_df: DataFrame,
-    builtin_metrics: Dict[str, int],  # pylint: disable=unused-argument
-) -> int:
+    builtin_metrics: Dict[str, float],  # pylint: disable=unused-argument
+) -> float:
     """
     Computes the weighted mean squared error (MSE) metric.
